style: ModalBasedOnMemeberCount 파일 분리
