+++ client/src/pages/EventPage/AdminPage/AdminPage.tsx
@@ -5,53 +5,9 @@ import StepList from '@components/StepList/StepList';
 import {useStepList} from '@hooks/useStepList/useStepList';
 import {requestGetEventName} from '@apis/request/event';
 import useEventId from '@hooks/useEventId/useEventId';
-
-import {SetActionListModal, SetInitialMemberListModal, SetAllMemberListModal} from '@components/Modal';
+import ModalBasedOnMemberCount from '@components/Modal/ModalBasedOnMemberCount/ModalBasedOnMemberCount';
 
 import {receiptStyle, titleAndListButtonContainerStyle} from './AdminPage.style';
-
-interface ModalBasedOnMemberCountProps {
-  memberNameList: string[];
-  isOpenBottomSheet: boolean;
-  isOpenAllMemberListButton: boolean;
-  setOrder: React.Dispatch<React.SetStateAction<number>>;
-  setIsOpenBottomSheet: React.Dispatch<React.SetStateAction<boolean>>;
-  setIsOpenAllMemberListButton: React.Dispatch<React.SetStateAction<boolean>>;
-}
-
-const ModalBasedOnMemberCount = ({
-  memberNameList,
-  isOpenBottomSheet,
-  isOpenAllMemberListButton,
-  setOrder,
-  setIsOpenBottomSheet,
-  setIsOpenAllMemberListButton,
-}: ModalBasedOnMemberCountProps) => {
-  if (isOpenAllMemberListButton) {
-    return (
-      <SetAllMemberListModal
-        setIsOpenBottomSheet={setIsOpenBottomSheet}
-        isOpenBottomSheet={isOpenBottomSheet}
-        setIsOpenAllMemberListButton={setIsOpenAllMemberListButton}
-      />
-    );
-  }
-  switch (memberNameList.length) {
-    case 0:
-      return (
-        <SetInitialMemberListModal setIsOpenBottomSheet={setIsOpenBottomSheet} isOpenBottomSheet={isOpenBottomSheet} />
-      );
-
-    default:
-      return (
-        <SetActionListModal
-          setOrder={setOrder}
-          setIsOpenBottomSheet={setIsOpenBottomSheet}
-          isOpenBottomSheet={isOpenBottomSheet}
-        />
-      );
-  }
-};
 
 const AdminPage = () => {
   const [isOpenFixedButtonBottomSheet, setIsOpenFixedBottomBottomSheet] = useState(false);
